+++ src/main/java/com/waitit/capstone/domain/admin/AdminMapper.java
@@ -2,8 +2,8 @@ package com.waitit.capstone.domain.admin;
 
 import com.waitit.capstone.domain.admin.dto.AllHostRequest;
 import com.waitit.capstone.domain.admin.dto.AllUserRequest;
-import com.waitit.capstone.domain.client.manager.Host;
-import com.waitit.capstone.domain.client.member.Entity.Member;
+import com.waitit.capstone.domain.manager.Host;
+import com.waitit.capstone.domain.member.Entity.Member;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
