+++ src/main/java/com/waitit/capstone/domain/admin/dto/AllUserRequest.java
@@ -1,7 +1,7 @@
 package com.waitit.capstone.domain.admin.dto;
 
 
-import com.waitit.capstone.domain.client.member.Entity.Gender;
+import com.waitit.capstone.domain.member.Entity.Gender;
 import lombok.Builder;
 import lombok.Getter;
