+++ src/main/java/com/waitit/capstone/domain/auth/dto/LoginRequest.java
@@ -1,0 +1,11 @@
+package com.waitit.capstone.domain.auth.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequest {
+    private String username;
+    private String password;
+}
