--- src/main/java/com/waitit/capstone/domain/client/manager/HostMapper.java
@@ -1,15 +1,0 @@
-package com.waitit.capstone.domain.client.manager;
-
-
-import com.waitit.capstone.domain.admin.dto.AllHostRequest;
-import com.waitit.capstone.domain.client.manager.dto.HostRequest;
-import com.waitit.capstone.domain.client.manager.dto.HostResponse;
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface HostMapper {
-
-    HostResponse hostToDto(Host host);
-
-    Host toEntity(HostRequest request);
-}
