--- src/main/java/com/waitit/capstone/domain/client/manager/HostRepository.java
@@ -1,9 +1,0 @@
-package com.waitit.capstone.domain.client.manager;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface HostRepository extends JpaRepository<Host,Long> {
-    Host findHostById(Long id);
-}
