--- src/main/java/com/waitit/capstone/domain/client/member/Entity/Gender.java
@@ -1,5 +1,0 @@
-package com.waitit.capstone.domain.client.member.Entity;
-
-public enum Gender {
-    MALE,FEMALE;
-}
