+++ src/main/java/com/waitit/capstone/domain/manager/HostRepository.java
@@ -1,0 +1,9 @@
+package com.waitit.capstone.domain.manager;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface HostRepository extends JpaRepository<Host,Long> {
+    Host findHostById(Long id);
+}
