+++ src/main/java/com/waitit/capstone/domain/manager/dto/HostResponse.java
@@ -1,0 +1,17 @@
+package com.waitit.capstone.domain.manager.dto;
+
+import java.time.LocalDateTime;
+
+public record HostResponse(
+        String imgUrl,
+        String hostName,
+        Integer maxPeople,
+        String hostManagerName,
+        String hostPhoneNumber,
+        Double latitude,
+        Double longitude,
+        String keyword,
+        String description,
+        LocalDateTime startTime,
+        LocalDateTime endTime) {
+}
