+++ src/main/java/com/waitit/capstone/domain/member/Entity/Gender.java
@@ -1,0 +1,5 @@
+package com.waitit.capstone.domain.member.Entity;
+
+public enum Gender {
+    MALE,FEMALE;
+}
