+++ src/main/java/com/waitit/capstone/domain/member/MemberRepository.java
@@ -1,0 +1,14 @@
+package com.waitit.capstone.domain.member;
+
+import com.waitit.capstone.domain.member.Entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member,Long> {
+    Boolean existsByNickname(String nickname);
+
+    Member findByPhoneNumber(String phoneNumber);
+
+    Member findMemberByPhoneNumber(String phoneNumber);
+
+    Member findMemberById(Long id);
+}
