+++ src/main/java/com/waitit/capstone/domain/member/MemberService.java
@@ -1,0 +1,26 @@
+package com.waitit.capstone.domain.member;
+
+import com.waitit.capstone.domain.admin.dto.UpdatedRequest;
+import com.waitit.capstone.domain.auth.dto.CustomUserDetails;
+import com.waitit.capstone.domain.member.Entity.Member;
+import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+
+
+    public Member updateMember(UpdatedRequest updatedRequest){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails customUserDetails = (CustomUserDetails) authentication.getPrincipal();
+        Member member = customUserDetails.getMember();
+        member.updateProfile(updatedRequest.getName(), updatedRequest.getNickName(), updatedRequest.getPassword());
+
+        return memberRepository.save(member);
+    }
+}
