+++ src/main/java/com/waitit/capstone/domain/message/SmsController.java
@@ -1,0 +1,23 @@
+package com.waitit.capstone.domain.message;
+
+import jakarta.validation.Valid;
+import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@AllArgsConstructor
+@RestController
+@RequestMapping("/sms")
+public class SmsController {
+
+    private final SmsService smsService;
+
+    @PostMapping("/send")
+    public ResponseEntity<?> SendSMS(@RequestBody @Valid SmsRequest smsRequest){
+        smsService.sendSms(smsRequest);
+        return ResponseEntity.ok("문자를 전송했습니다.");
+    }
+}
