+++ src/main/java/com/waitit/capstone/global/security/jwt/LoginFilter.java
@@ -1,18 +1,15 @@
 package com.waitit.capstone.global.security.jwt;
 
-import com.waitit.capstone.domain.client.auth.dto.LoginRequest;
+import com.waitit.capstone.domain.auth.dto.LoginRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.waitit.capstone.domain.client.auth.service.RefreshTokenService;
+import com.waitit.capstone.domain.auth.service.RefreshTokenService;
 import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletInputStream;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.util.Collection;
-import java.util.Iterator;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -20,7 +17,6 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.util.StreamUtils;
