+++ src/main/resources/application.properties
@@ -17,3 +17,9 @@ spring.jwt.secret=${SPRING_JWT_SECRET}
 # Redis
 spring.data.redis.host=${SPRING_REDIS_HOST}
 spring.data.redis.port=${SPRING_REDIS_PORT}
+
+
+# SMS
+coolsms.apikey=${COOL_API_KEY}
+coolsms.apisecret=${COOL_API_SECRET}
+coolsms.fromnumber=${COOL_PHONE_NUMBER}
